GCD FADC gain scale factor and HV Pulse checks (#34)
* Create GCD_ScaleFactor_Check.py

Checks the FADC Scale Factor in the given GCD file compared to the expected scale factor. Plots the distribution of the gcd scale factor (for each DOM) compared to the nominal ones. 

Also gives a text file out with the GCD and nominal scale factors for each DOM

* Create HV_Pulse_Check.py

checks which DOMs had HV in a run and then looks for pulses in those DOMs from the entire run. Outputs a file listing doms that did NOT have a pulse (but had HV on)